split rpc implementation out
