+++ light-client-bin/Cargo.toml
@@ -20,6 +20,7 @@ ckb-traits = "1"
 ckb-types = "1"
 
 ckb-light-client-lib = {path = "../light-client-lib", default-features = false}
+ckb-light-client-rpc = {path = "../light-client-rpc", default-features = false}
 clap = {version = "4", features = ["cargo"]}
 ctrlc = {version = "3.2.1", features = ["termination"]}
 jsonrpc-core = "18.0"
@@ -51,8 +52,8 @@ tempfile = "3.0"
 default = ["rocksdb"]
 march-native = ["rocksdb/march-native"]
 portable = ["rocksdb/portable"]
-rocksdb = ["dep:rocksdb", "ckb-light-client-lib/rocksdb"]
-rusqlite = ["ckb-light-client-lib/rusqlite", "dep:rusqlite"]
+rocksdb = ["dep:rocksdb", "ckb-light-client-lib/rocksdb", "ckb-light-client-rpc/rocksdb"]
+rusqlite = ["ckb-light-client-lib/rusqlite", "ckb-light-client-rpc/rusqlite", "dep:rusqlite"]
 # [profile.release]
 # overflow-checks = true
